+++ app/src/main/java/com/techcourse/DispatcherServletInitializer.java
@@ -6,8 +6,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.interface21.web.WebApplicationInitializer;
-import com.interface21.webmvc.servlet.mvc.tobe.adapter.HandlerAdapterContainer;
-import com.interface21.webmvc.servlet.mvc.tobe.mapping.HandlerMappingContainer;
+import com.interface21.webmvc.servlet.mvc.adapter.HandlerAdapterContainer;
+import com.interface21.webmvc.servlet.mvc.mapping.HandlerMappingContainer;
 
 /**
  * Base class for {@link WebApplicationInitializer}
